# Conflicts:
#	app/src/main/java/com/danilkharytonov/retrofitroomrepository/presentation/base/BaseFragment.kt
#	app/src/main/java/com/danilkharytonov/retrofitroomrepository/presentation/detail_user/UserDetailFragment.kt
#	app/src/main/java/com/danilkharytonov/retrofitroomrepository/presentation/user_list/UserListFragment.kt

+++ app/src/main/java/com/danilkharytonov/retrofitroomrepository/domain/use_cases/user_list/GetUsersUseCase.kt
@@ -7,16 +7,20 @@ import com.danilkharytonov.retrofitroomrepository.domain.repository.UserStorageR
 import java.net.UnknownHostException
 import javax.inject.Inject
 
-class GetAllUsersFromApiUseCase @Inject constructor(
+class GetUsersUseCase @Inject constructor(
     private val apiRepository: UserApiRepository,
     private val dbRepository: UserDBRepository,
     private val storageRepository: UserStorageRepository
 ) {
+    private var isFirstTime = true
     suspend fun execute(results: Int): List<User> {
         return try {
             val users = apiRepository.getAllUsersFromApi(results)
-            dbRepository.deleteUsersFromDB()
             dbRepository.insertUsersToDB(users)
+            if (!isFirstTime) {
+                dbRepository.deleteUsersFromDB()
+                isFirstTime = false
+            }
             storageRepository.saveUserImagesInStorage(users)
             users
         } catch (e: UnknownHostException) {
